% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_BuildBackend.R
\name{build_backend}
\alias{build_backend}
\title{An optional function to build the environment for fastrerandomize. Builds a conda environment in which jax and np are installed. Users can also create a conda environment where jax and np are installed themselves.}
\usage{
build_backend(conda_env = "fastrerandomize", conda = "auto")
}
\arguments{
\item{conda_env}{(default = \code{"fastrerandomize"}) Name of the conda environment in which to place the backends.}

\item{conda}{(default = \code{auto}) The path to a conda executable. Using \code{"auto"} allows reticulate to attempt to automatically find an appropriate conda binary.}
}
\value{
Builds the computational environment for \code{fastrerandomize}. This function requires an Internet connection.
You may find out a list of conda Python paths via: \code{Sys.which("python")}
}
\description{
An optional function to build the environment for fastrerandomize. Builds a conda environment in which jax and np are installed. Users can also create a conda environment where jax and np are installed themselves.
}
\examples{
# For a tutorial, see
# github.com/cjerzak/fastrerandomize-software/

}
