% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_BuildBackend.R
\name{BuildBackend}
\alias{BuildBackend}
\title{Build the environment for fastrerandomize. Builds a conda environment in which jax and np are installed.}
\usage{
BuildBackend(conda_env = "fastrerandomize", conda = "auto", tryMetal = T)
}
\arguments{
\item{conda_env}{(default = \code{"fastrerandomize"}) Name of the conda environment in which to place the backends.}

\item{conda}{(default = \code{auto}) The path to a conda executable. Using \code{"auto"} allows reticulate to attempt to automatically find an appropriate conda binary.}
}
\value{
Builds the computational environment for \code{fastrerandomize}. This function requires an Internet connection.
You may find out a list of conda Python paths via: \code{system("which python")}
}
\description{
Build the environment for fastrerandomize. Builds a conda environment in which jax and np are installed.
}
\examples{
# For a tutorial, see
# github.com/cjerzak/fastrerandomize-software/

}
