% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_HelperFxns.R
\name{check_jax_availability}
\alias{check_jax_availability}
\title{Check if 'Python' and 'JAX' are available}
\usage{
check_jax_availability(conda_env = "fastrerandomize", conda = "auto")
}
\arguments{
\item{conda_env}{A character string specifying the name of the conda environment. 
Default is `"fastrerandomize"`.}

\item{conda}{The path to a conda executable, or `"auto"`. Default is `"auto"`.}
}
\value{
Returns `TRUE` (invisibly) if both 'Python' and 'JAX' are available; otherwise returns `NULL`.
}
\description{
This function checks if 'Python' and 'JAX' can be accessed via `reticulate`. If not,
it returns `NULL` and prints a message suggesting to run `build_backend()`.
}
\examples{
\dontrun{
  check_jax_availability()
}

}
