% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_BaseRFxns.R
\name{compute_diff_at_tau_for_oneW_R}
\alias{compute_diff_at_tau_for_oneW_R}
\title{Compute potential outcome difference in means for a single assignment
  under a hypothesized tau in base R}
\usage{
compute_diff_at_tau_for_oneW_R(Wprime, obsY, obsW, tau)
}
\arguments{
\item{Wprime}{A 0/1 assignment vector for which to compute the diff in means.}

\item{obsY}{Observed outcome vector.}

\item{obsW}{Observed assignment vector.}

\item{tau}{The hypothesized true effect for the shift in outcomes under treatment.}
}
\value{
Scalar difference in means for the assignment Wprime.
}
\description{
Compute potential outcome difference in means for a single assignment
  under a hypothesized tau in base R
}
