% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_BaseRFxns.R
\name{diff_in_means_R}
\alias{diff_in_means_R}
\title{Simple difference in means in base R}
\usage{
diff_in_means_R(Y, W)
}
\arguments{
\item{Y}{Numeric outcome vector.}

\item{W}{0/1 treatment assignment vector.}
}
\value{
Scalar difference in means: mean(Y|W=1) - mean(Y|W=0).
}
\description{
Simple difference in means in base R
}
