% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_S3Constructors.R
\name{fastrerandomize_class}
\alias{fastrerandomize_class}
\title{Constructor for fastrerandomize randomizations}
\usage{
fastrerandomize_class(
  randomizations,
  balance = NULL,
  fastrr_env = NULL,
  call = NULL
)
}
\arguments{
\item{randomizations}{A matrix or array where each row (or slice) represents one 
randomization.}

\item{balance}{A numeric vector or similar object holding balance statistics 
for each randomization, or \code{NULL} if not applicable.}

\item{fastrr_env}{Associated \code{fastrr_env} environment.}

\item{call}{The function call, if you wish to store it for reference (optional).}
}
\value{
An object of class \code{fastrerandomize_randomizations}.
}
\description{
Create an S3 object of class \code{fastrerandomize_randomizations} that stores
the randomizations (and optionally balance statistics) generated by 
functions such as \code{\link{generate_randomizations}}.
}
