% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_GenerateData.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data()
}
\value{
A list consiting of \itemize{
\item \code{pval} A p-value.
}
}
\description{
Generate data
}
\section{References}{

\itemize{
\item
}
}

\examples{
# For a tutorial, see
# github.com/cjerzak/fastrerandomization

}
