% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_BaseRFxnVersion.R
\name{hotellingT2_R}
\alias{hotellingT2_R}
\title{Compute Hotelling's T^2 in base R}
\usage{
hotellingT2_R(X, W)
}
\arguments{
\item{X}{A numeric matrix of covariates (observations in rows).}

\item{W}{A 0/1 treatment assignment vector of the same length as nrow(X).}
}
\value{
A numeric scalar: the Hotelling's T^2 for that assignment.
}
\description{
This function provides a base R implementation of Hotelling's T^2
balance measure, renamed with `_R` for clarity that it is the R-based
analog to the JAX version in fastrerandomize.
}
