% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_InitializeJAX.R
\name{InitializeJax}
\alias{InitializeJax}
\title{Initialize JAX}
\usage{
InitializeJax()
}
\value{
A list consiting of \itemize{
\item \code{pval} A p-value.
}
}
\description{
Initialize JAX
}
\section{References}{

\itemize{
\item
}
}

\examples{
# For a tutorial, see
# github.com/cjerzak/fastrerandomization

}
