% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_InitializeJAXfxns.R
\name{InitializeJAXFxns}
\alias{InitializeJAXFxns}
\title{Initialize JAX functions}
\usage{
InitializeJAXFxns()
}
\arguments{
\item{...}{This function takes no arguments}
}
\value{
This function should be run just after \code{fastrerandomize::InitializeJax}.This sets up the internal JAX functions for rerandomization analysis.
}
\description{
Initialize JAX functions
}
\examples{
# For tutorials, see
# github.com/cjerzak/fastrerandomization-software

}
