% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_S3Constructors.R
\name{plot.fastrerandomize_randomizations}
\alias{plot.fastrerandomize_randomizations}
\title{Plot method for fastrerandomize_randomizations objects}
\usage{
\method{plot}{fastrerandomize_randomizations}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fastrerandomize_randomizations}.}

\item{...}{Further arguments passed to base \code{\link{plot}} functions.}
}
\description{
Plot method for fastrerandomize_randomizations objects
}
