% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_S3Constructors.R
\name{plot.fastrerandomize_test}
\alias{plot.fastrerandomize_test}
\title{Plot method for fastrerandomize_test objects}
\usage{
\method{plot}{fastrerandomize_test}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fastrerandomize_test}.}

\item{...}{Further graphical parameters passed to \code{\link{plot}}.}
}
\value{
No output returned. Performs side effect of plotting \code{fastrerandomize_test} class objects.
}
\description{
Plots a simple visualization of the observed effect and the 
fiducial interval (if present) on a horizontal axis.
}
