% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_S3Constructors.R
\name{print.fastrerandomize_randomizations}
\alias{print.fastrerandomize_randomizations}
\title{Print method for fastrerandomize_randomizations objects}
\usage{
\method{print}{fastrerandomize_randomizations}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fastrerandomize_instance}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Prints an object of class \code{fastrerandomize_randomizations}.
}
\description{
Print method for fastrerandomize_randomizations objects
}
