% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_QJEData.R
\docType{data}
\name{QJEData}
\alias{QJEData}
\title{QJEData: Agricultural Treatment Experiment Data}
\format{
A data frame with 968 rows and 2161 column. 
  \describe{
    \item{rows}{Indicate units}
    \item{columns}{Indicate RCT data}
  }
}
\source{
Burchardi, K.B., Ghatak, M., & Johanssen, A. (2019). 
Moral hazard: Experimental evidence from tenancy contracts.
_The Quarterly Journal of Economics_, 134(1), 281-347.
}
\usage{
data(QJEData)
}
\description{
Data from a field experiment studying moral hazard in tenancy contracts in agriculture.
}
