% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_S3Constructors.R
\name{summary.fastrerandomize_randomizations}
\alias{summary.fastrerandomize_randomizations}
\title{Summary method for fastrerandomize_randomizations objects}
\usage{
\method{summary}{fastrerandomize_randomizations}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fastrerandomize_randomizations}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list with summary statistics, printed by default.
}
\description{
Summary method for fastrerandomize_randomizations objects
}
