% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_S3Constructors.R
\name{summary.fastrerandomize_test}
\alias{summary.fastrerandomize_test}
\title{Summary method for fastrerandomize_test objects}
\usage{
\method{summary}{fastrerandomize_test}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fastrerandomize_test}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Returns an (invisible) list with a summary of \code{fastrerandomize_test} class objects.
}
\description{
Summary method for fastrerandomize_test objects
}
