% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_GenerateData.R
\name{GenerateCausalData}
\alias{GenerateCausalData}
\title{This function generates simulated causal data based on specified parameters.}
\usage{
GenerateCausalData(n_units)
}
\arguments{
\item{n_units}{A numeric value specifying the total number of units in the sample.}

\item{proportion_treated}{A numeric value between 0 and 1 indicating the proportion of units that receive treatment.}

\item{k_covars}{A numeric value indicating the number of covariates to be generated.}

\item{rho}{A numeric value representing the correlation coefficient.}

\item{SD_inherent}{A numeric value indicating the standard deviation inherent to the data.}

\item{treatment_effect_mean}{A numeric value representing the mean of the treatment effect.}

\item{treatment_effect_SD}{A numeric value indicating the standard deviation of the treatment effect.}

\item{covariates_SD}{A numeric value or vector specifying the standard deviation of the covariates.}

\item{Y0_coefficients}{An optional numeric vector specifying the coefficients for the control outcome model. If not provided, the function assumes a NULL value.}

\item{Y1_coefficients}{An optional numeric vector specifying the coefficients for the treated outcome model. If not provided, the function assumes a NULL value.}
}
\value{
A list consisting of \itemize{
\item \code{data_matrix} A data frame containing the simulated covariates and outcomes for both control (Y0) and treatment (Y1) groups.
\item \code{Y0_coefficients} A numeric vector representing the coefficients used for the control outcome model.
\item \code{Y1_coefficients} A numeric vector representing the coefficients used for the treated outcome model.
}
}
\description{
This function generates simulated causal data based on specified parameters.
}
\examples{
# For a tutorial, see
# github.com/cjerzak/fastrerandomization-software

}
