% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_RandomizationTest.R
\name{GenerateRandomizations}
\alias{GenerateRandomizations}
\title{Fast generation of all possible complete randomizations given target number of experimental units.}
\usage{
GenerateRandomizations(
  n_units,
  n_treated,
  X = NULL,
  randomization_accept_prob = 1
)
}
\arguments{
\item{n_units}{A integer specifying total number of experimental units.}

\item{n_treated}{An integer specifying total number of treated units.}
}
\value{
A JAX array containing all possible complete randomizations.
}
\description{
Fast generation of all possible complete randomizations given target number of experimental units.
}
\examples{
# For a tutorial, see
# github.com/cjerzak/fastrerandomization-software

}
