% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_InitializeJAX.R
\name{InitializeJAX}
\alias{InitializeJAX}
\title{Initialize JAX}
\usage{
InitializeJax(conda_env, conda_env_required)
}
\arguments{
\item{conda_env}{An optioanl character string representing the conda environment to activate. A version of JAX should live in that environment. If NULL, we look in the default Python environment for JAX.}

\item{conda_env_required}{A logical representing whether to force use the specified conda environment. Used only if \code{conda_env} specified.}
}
\value{
This function initializes a JAX-containing conda environment as specified by \code{conda_env}. This function must be run before any others in \code{fastrerandomize}.
}
\description{
Initialize JAX
}
\examples{
# For a tutorial, see
# github.com/cjerzak/fastrerandomization-software

}
