% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_InitializeJAXfxns.R
\name{InitializeJAXFxns}
\alias{InitializeJAXFxns}
\title{Initialize JAX}
\usage{
InitializeJAXFxns()
}
\value{
A list consiting of \itemize{
\item
}
}
\description{
Initialize JAX
}
\examples{
# For tutorials, see
# github.com/cjerzak/fastrerandomization-software

}
