% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_QJEData.R
\docType{data}
\name{QJEData}
\alias{QJEData}
\title{QJEData: Agricultural Treatment Experiment Data}
\format{
A data frame with 968 rows and the following columns:
  \describe{
    \item{column1}{Description of first column}
    \item{column2}{Description of second column}
  }
}
\source{
Burchardi, K.B., Ghatak, M., & Johanssen, A. (2019). 
Moral hazard: Experimental evidence from tenancy contracts.
The Quarterly Journal of Economics, 134(1), 281-347.
}
\usage{
data(QJEData)
}
\description{
Data from a field experiment studying moral hazard in tenancy contracts in agriculture.
}
