% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_QJEData.R
\docType{data}
\name{QJEData}
\alias{QJEData}
\title{QJEData}
\format{
A data frame with 968 rows and many columns containing treatment data for a Quarterly Journal of Economics experiment on agriculture.
}
\source{
Burchardi, Konrad B et al. (2019). "Moral hazard: Experimental evidence from tenancy contracts."
In: The Quarterly Journal of Economics 134.1, pp. 281–347
}
\usage{
QJEData
}
\description{
The dataset originates from the study "Moral hazard: Experimental evidence from tenancy contracts" by
Burchardi, Konrad B et al., published in "The Quarterly Journal of Economics" in 2019 (Volume 134, Issue 1, Pages 281-347).
}
\keyword{datasets}
