% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_RandomizationTest.R
\name{randomization_test}
\alias{randomization_test}
\title{Fast randomization test}
\usage{
randomization_test(X, ...)
}
\arguments{
\item{obsY}{A numeric vector containing observed outcomes.}

\item{obsW}{A numeric vector where \code{0}'s correspond to control units and \code{1}'s to treated units.}
}
\value{
A list consiting of \itemize{
\item \code{pval} A p-value.
}
}
\description{
Fast randomization test
}
\section{References}{

\itemize{
\item
}
}

\examples{
# For a tutorial, see
# github.com/cjerzak/fastrerandomization

}
