% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRR_RandomizationTest.R
\name{randomization_test}
\alias{randomization_test}
\title{Fast randomization test}
\usage{
randomization_test(
  obsW = NULL,
  obsY = NULL,
  X = NULL,
  alpha = 0.05,
  candidate_randomizations = NULL,
  candidate_randomizations_array = NULL,
  n0_array = NULL,
  n1_array = NULL,
  randomization_accept_prob = 1,
  findFI = F,
  c_initial = 2,
  max_draws = 10^6,
  batch_size = 10^5,
  randomization_type = "monte_carlo",
  approximate_inv = TRUE,
  file = NULL,
  conda_env = "fastrerandomize",
  conda_env_required = T
)
}
\arguments{
\item{obsW}{A numeric vector where \code{0}'s correspond to control units and \code{1}'s to treated units.}

\item{obsY}{An optional numeric vector of observed outcomes. If not provided, the function assumes a NULL value.}

\item{X}{A numeric matrix of covariates.}

\item{alpha}{The significance level for the test. Default is \code{0.05}.}

\item{candidate_randomizations}{A numeric matrix of candidate randomizations.}

\item{candidate_randomizations_array}{An optional JAX array of candidate randomizations. If not provided, the function coerces \code{candidate_randomizations} into a JAX array.}

\item{n0_array}{An optional array specifying the number of control units.}

\item{n1_array}{An optional array specifying the number of treated units.}

\item{randomization_accept_prob}{An numeric scalar or vector of probabilities for accepting each randomization.}

\item{findFI}{A logical value indicating whether to find the fiducial interval. Default is FALSE.}

\item{c_initial}{A numeric value representing the initial criterion for the randomization. Default is \code{2}.}

\item{true_treatment_effect}{An optional numeric value specifying the true treatment effect. Default is NULL.}
}
\value{
A list consisting of \itemize{
\item \code{p_value} A numeric value or vector representing the p-value of the test (or the expected p-value under the prior structure specified in the function inputs).
\item \code{FI} A numeric vector representing the fiducial interval if \code{findFI=T}.
\item \code{tau_obs} A numeric value or vector representing the estimated treatment effect(s)
}
}
\description{
Fast randomization test
}
\section{References}{

\itemize{
\item
}
}

\examples{
# For a tutorial, see
# github.com/cjerzak/fastrerandomization-software

}
